Refactor: view  분리
